ai-service/requirements.txt
# AI Service Dependencies
# Python 3.11 or 3.12 recommended (Python 3.14 has protobuf compatibility issues)

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Google AI
google-genai>=0.3.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Data Validation
# Using newer pydantic that supports Python 3.14
pydantic>=2.9.0,<3.0.0

# Image Processing
# Using latest Pillow version that supports Python 3.14
Pillow>=10.4.0,<11.0.0

# Utilities
python-dateutil==2.8.2

# Logging
structlog==24.1.0
